+++ example/lib/main.dart
@@ -34,6 +34,8 @@ import 'package:example/pages/switch_form_field.dart';
 import 'package:example/pages/table.dart';
 import 'package:example/pages/tabs.dart';
 import 'package:example/pages/portal.dart';
+import 'package:example/pages/textarea.dart';
+import 'package:example/pages/textarea_form_field.dart';
 import 'package:example/pages/time_picker.dart';
 import 'package:example/pages/time_picker_form_field.dart';
 import 'package:example/pages/toast.dart';
@@ -85,6 +87,8 @@ final routes = <String, WidgetBuilder>{
   '/switch-form-field': (_) => const SwitchFormFieldPage(),
   '/table': (_) => const TablePage(),
   '/tabs': (_) => const TabsPage(),
+  '/textarea': (_) => const TextareaPage(),
+  '/textarea-form-field': (_) => const TextareaFormFieldPage(),
   '/time-picker': (_) => const TimePickerPage(),
   '/time-picker-form-field': (_) => const TimePickerFormFieldPage(),
   '/toast': (_) => const ToastPage(),
@@ -107,15 +111,20 @@ class App extends StatelessWidget {
         // Custom App example
         // return ShadApp.custom(
         //   themeMode: themeMode,
-        //   appBuilder: (context, theme) => GetMaterialApp(
-        //     routes: routes,
-        //     themeMode: themeMode,
-        //     theme: theme,
-        //     home: const MainPage(),
-        //     builder: (context, child) {
-        //       return ShadToaster(child: child!);
-        //     },
+        //   darkTheme: ShadThemeData(
+        //     brightness: Brightness.dark,
+        //     colorScheme: const ShadSlateColorScheme.dark(),
         //   ),
+        //   appBuilder: (context) {
+        //     return MaterialApp(
+        //       routes: routes,
+        //       theme: Theme.of(context),
+        //       home: const MainPage(),
+        //       builder: (context, child) {
+        //         return ShadAppBuilder(child: child!);
+        //       },
+        //     );
+        //   },
         // );
         return ShadApp(
           debugShowCheckedModeBanner: false,